pyproject.toml
[project]
name = "aminder-personal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "openmemory-py>=1.2.3",
    "pyyaml>=6.0.3",
]
